rolling back seldon example since it's not fully tested and can't be run with yaml yet
