+++ examples/seldon_deployment/README.md
@@ -613,4 +613,4 @@ rm -rf zenml_examples
 Our docs regarding the seldon deployment integration can be found [here](https://docs.zenml.io/advanced-guide/continuous-training-and-deployment).
 
 If you want to learn more about deployment in ZenML in general or about how to build your own deployer steps in ZenML
-check out our [docs](https://docs.zenml.io/extending-zenml/model-deployers).
+check out our [docs](https://docs.zenml.io/extending-zenml/model-deployer).
