--- examples/seldon_deployment/continuous_deployment_config_sklearn.yaml
@@ -1,31 +1,0 @@
-name: continuous_deployment_pipeline
-steps:
-  importer_mnist:
-    source:
-      file: steps/importer_mnist/importer_step
-      name: importer_mnist
-  normalizer:
-    source:
-      file: steps/normalizer/normalizer_step
-      name: normalizer
-  sklearn_trainer:
-    source:
-      file: steps/sklearn_trainer/sklearn_trainer_step
-      name: sklearn_trainer
-  sklearn_evaluator:
-    source:
-      file: steps/sklearn_evaluator/sklearn_evaluator_step
-      name: sklearn_evaluator
-  deployment_trigger:
-    source:
-      file: steps/deployment_trigger/deployment_trigger_step
-      name: deployment_trigger
-    parameters:
-      min_accuracy: 0.8
-  seldon_model_deployer_step:
-    source:
-      file: steps/model_deployer/model_deployer_step.py
-      name: seldon_model_deployer_step
-    parameters:
-      model_name: "mnist"
-      implementation: "SKLEARN_SERVER"
