--- examples/seldon_deployment/continuous_deployment_config_tensorflow.yaml
@@ -1,31 +1,0 @@
-name: continuous_deployment_pipeline
-steps:
-  importer_mnist:
-    source:
-      file: steps/importer_mnist/importer_step
-      name: importer_mnist
-  normalizer:
-    source:
-      file: steps/normalizer/normalizer_step
-      name: normalizer
-  tf_trainer:
-    source:
-      file: steps/tf_trainer/tf_trainer_step
-      name: tf_trainer
-  tf_evaluator:
-    source:
-      file: steps/tf_evaluator/tf_evaluator_step
-      name: tf_evaluator
-  deployment_trigger:
-    source:
-      file: steps/deployment_trigger/deployment_trigger_step
-      name: deployment_trigger
-    parameters:
-      min_accuracy: 0.8
-  seldon_model_deployer_step:
-    source:
-      file: steps/model_deployer/model_deployer_step.py
-      name: seldon_model_deployer_step
-    parameters:
-      model_name: "mnist"
-      implementation: "TENSORFLOW_SERVER"
