--- examples/seldon_deployment/pipelines/__init__.py
@@ -1,22 +1,0 @@
-#  Copyright (c) ZenML GmbH 2022. All Rights Reserved.
-#
-#  Licensed under the Apache License, Version 2.0 (the "License");
-#  you may not use this file except in compliance with the License.
-#  You may obtain a copy of the License at:
-#
-#       https://www.apache.org/licenses/LICENSE-2.0
-#
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
-#  or implied. See the License for the specific language governing
-#  permissions and limitations under the License.
-from .continuous_deployment.continous_deployment_pipeline import (
-    continuous_deployment_pipeline,
-)
-from .inference.inference_pipeline import inference_pipeline
-
-__all__ = [
-    "continuous_deployment_pipeline",
-    "inference_pipeline",
-]
