+++ examples/seldon_deployment/run.py
@@ -14,42 +14,37 @@
 from typing import cast
 
 import click
-from rich import print
-from steps.deployment_trigger.deployment_trigger_step import (
+from pipeline import (
     DeploymentTriggerConfig,
-    deployment_trigger,
-)
-from steps.dynamic_importer.dynamic_importer_step import dynamic_importer
-from steps.importer_mnist.importer_mnist_step import importer_mnist
-from steps.model_deployer.model_deployer_step import seldon_model_deployer_step
-from steps.normalizer.normalizer_step import normalizer
-from steps.prediction_service_loader.prediction_service_loader_step import (
     SeldonDeploymentLoaderStepConfig,
-    prediction_service_loader,
-)
-from steps.predictor.predictor_step import predictor
-from steps.sklearn_evaluator.sklearn_evaluator_step import sklearn_evaluator
-from steps.sklearn_predict_preprocessor.sklearn_predict_preprocessor_step import (
-    sklearn_predict_preprocessor,
-)
-from steps.sklearn_trainer.sklearn_trainer_step import (
     SklearnTrainerConfig,
+    TensorflowTrainerConfig,
+    continuous_deployment_pipeline,
+    deployment_trigger,
+    dynamic_importer,
+    importer_mnist,
+    inference_pipeline,
+    normalizer,
+    prediction_service_loader,
+    predictor,
+    sklearn_evaluator,
+    sklearn_predict_preprocessor,
     sklearn_trainer,
-)
-from steps.tf_evaluator.tf_evaluator_step import tf_evaluator
-from steps.tf_predict_preprocessor.tf_predict_preprocessor_step import (
+    tf_evaluator,
     tf_predict_preprocessor,
+    tf_trainer,
 )
-from steps.tf_trainer.tf_trainer_step import TensorflowTrainerConfig, tf_trainer
+from rich import print
 
 from zenml.integrations.seldon.model_deployers import SeldonModelDeployer
 from zenml.integrations.seldon.services import (
     SeldonDeploymentConfig,
     SeldonDeploymentService,
 )
-from zenml.integrations.seldon.steps import SeldonDeployerStepConfig
-
-from .pipelines import continuous_deployment_pipeline, inference_pipeline
+from zenml.integrations.seldon.steps import (
+    SeldonDeployerStepConfig,
+    seldon_model_deployer_step,
+)
 
 DEPLOY = "deploy"
 PREDICT = "predict"
