--- examples/seldon_deployment/steps/deployment_trigger/deployment_trigger_step.py
@@ -1,31 +1,0 @@
-#  Copyright (c) ZenML GmbH 2022. All Rights Reserved.
-#
-#  Licensed under the Apache License, Version 2.0 (the "License");
-#  you may not use this file except in compliance with the License.
-#  You may obtain a copy of the License at:
-#
-#       https://www.apache.org/licenses/LICENSE-2.0
-#
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
-#  or implied. See the License for the specific language governing
-#  permissions and limitations under the License.
-from zenml.steps import BaseStepConfig, step
-
-
-class DeploymentTriggerConfig(BaseStepConfig):
-    """Parameters that are used to trigger the deployment"""
-
-    min_accuracy: float
-
-
-@step
-def deployment_trigger(
-    accuracy: float,
-    config: DeploymentTriggerConfig,
-) -> bool:
-    """Implements a simple model deployment trigger that looks at the
-    input model accuracy and decides if it is good enough to deploy"""
-
-    return accuracy > config.min_accuracy
