--- examples/seldon_deployment/steps/predictor/predictor_step.py
@@ -1,32 +1,0 @@
-#  Copyright (c) ZenML GmbH 2022. All Rights Reserved.
-#
-#  Licensed under the Apache License, Version 2.0 (the "License");
-#  you may not use this file except in compliance with the License.
-#  You may obtain a copy of the License at:
-#
-#       http://www.apache.org/licenses/LICENSE-2.0
-#
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
-#  or implied. See the License for the specific language governing
-#  permissions and limitations under the License.
-import numpy as np
-from rich import print as rich_print
-
-from zenml.integrations.seldon.services import SeldonDeploymentService
-from zenml.steps import Output, step
-
-
-@step
-def predictor(
-    service: SeldonDeploymentService,
-    data: np.ndarray,
-) -> Output(predictions=np.ndarray):
-    """Run a inference request against a prediction service"""
-
-    service.start(timeout=120)  # should be a NOP if already started
-    prediction = service.predict(data)
-    prediction = prediction.argmax(axis=-1)
-    rich_print("Prediction: ", prediction)
-    return prediction
