--- examples/seldon_deployment/steps/tf_predict_preprocessor/tf_predict_preprocessor_step.py
@@ -1,23 +1,0 @@
-#  Copyright (c) ZenML GmbH 2022. All Rights Reserved.
-#
-#  Licensed under the Apache License, Version 2.0 (the "License");
-#  you may not use this file except in compliance with the License.
-#  You may obtain a copy of the License at:
-#
-#       http://www.apache.org/licenses/LICENSE-2.0
-#
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
-#  or implied. See the License for the specific language governing
-#  permissions and limitations under the License.
-import numpy as np  # type: ignore [import]
-
-from zenml.steps import Output, step
-
-
-@step
-def tf_predict_preprocessor(input: np.ndarray) -> Output(data=np.ndarray):
-    """Prepares the data for inference."""
-    input = input / 255.0
-    return input
